moved flattenBy to utils
